requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose
passlib
python-multipart
google-genai
bcrypt